Move EvmRuntimeConfig into config.rs
