+++ crates/module-system/module-implementations/sov-evm/src/config.rs
@@ -63,6 +63,29 @@ impl Default for EvmGenesisConfig {
     }
 }
 
+/// Runtime configuration for EVM execution
+#[derive(Debug, Clone, PartialEq, serde::Deserialize, serde::Serialize)]
+pub struct EvmRuntimeConfig {
+    /// Core chain parameters
+    pub chain_spec: crate::EvmChainSpec,
+    /// Sorted hard fork schedule for efficient runtime lookup
+    /// (block number, fork ID) ordered by block number
+    pub hardforks: Vec<(u64, SpecId)>,
+}
+
+impl Default for EvmRuntimeConfig {
+    fn default() -> EvmRuntimeConfig {
+        let chain_spec = crate::EvmChainSpec::default();
+        // Clone hardforks from chain_spec for runtime use
+        let hardforks = chain_spec.hardforks.clone();
+
+        EvmRuntimeConfig {
+            chain_spec,
+            hardforks,
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use std::str::FromStr;
